added tests for moraldayx classes
